requirements.txt
django==4.1.2
djangorestframework==3.14.0
python-dotenv==0.21.0
pillow==9.2.0
